📦 Update utils and middleware configs

+++ middleware/session/session.go
@@ -10,11 +10,11 @@ import (
 )
 
 type Session struct {
-	ctx    *fiber.Ctx
-	config *Store
-	db     *db
-	id     string
-	fresh  bool
+	id     string     // session id
+	fresh  bool       // if new session
+	ctx    *fiber.Ctx // fiber context
+	config *Store     // store configuration
+	data   *data      // key value data
 }
 
 var sessionPool = sync.Pool{
@@ -25,19 +25,20 @@ var sessionPool = sync.Pool{
 
 func acquireSession() *Session {
 	s := sessionPool.Get().(*Session)
-	s.db = new(db)
+	if s.data == nil {
+		s.data = new(data)
+	}
 	s.fresh = true
 	return s
 }
 
 func releaseSession(s *Session) {
+	s.id = ""
 	s.ctx = nil
 	s.config = nil
-	if s.db != nil {
-		s.db.Reset()
+	if s.data != nil {
+		s.data.Reset()
 	}
-	s.id = ""
-	s.fresh = true
 	sessionPool.Put(s)
 }
 
@@ -53,25 +54,42 @@ func (s *Session) ID() string {
 
 // Get will return the value
 func (s *Session) Get(key string) interface{} {
-	return s.db.Get(key)
+	// Better safe than sorry
+	if s.data == nil {
+		return nil
+	}
+	return s.data.Get(key)
 }
 
 // Set will update or create a new key value
 func (s *Session) Set(key string, val interface{}) {
-	s.db.Set(key, val)
+	// Better safe than sorry
+	if s.data == nil {
+		return
+	}
+	s.data.Set(key, val)
 }
 
 // Delete will delete the value
 func (s *Session) Delete(key string) {
-	s.db.Delete(key)
+	// Better safe than sorry
+	if s.data == nil {
+		return
+	}
+	s.data.Delete(key)
 }
 
 // Destroy will delete the session from Storage and expire session cookie
 func (s *Session) Destroy() error {
-	// Reset local data
-	s.db.Reset()
+	// Better safe than sorry
+	if s.data == nil {
+		return nil
+	}
 
-	// Delete data from storage
+	// Reset local data
+	s.data.Reset()
+
+	// Use external Storage if exist
 	if err := s.config.Storage.Delete(s.id); err != nil {
 		return err
 	}
@@ -88,6 +106,7 @@ func (s *Session) Regenerate() error {
 	if err := s.config.Storage.Delete(s.id); err != nil {
 		return err
 	}
+
 	// Create new ID
 	s.id = s.config.KeyGenerator()
 
@@ -96,13 +115,18 @@ func (s *Session) Regenerate() error {
 
 // Save will update the storage and client cookie
 func (s *Session) Save() error {
-	// Don't save to Storage if no data is available
-	if s.db.Len() <= 0 {
+	// Better safe than sorry
+	if s.data == nil {
 		return nil
 	}
 
-	// Convert book to bytes
-	data, err := s.db.MarshalMsg(nil)
+	// Don't save to Storage if no data is available
+	if s.data.Len() <= 0 {
+		return nil
+	}
+
+	// Convert data to bytes
+	data, err := s.data.MarshalMsg(nil)
 	if err != nil {
 		return err
 	}
@@ -115,7 +139,8 @@ func (s *Session) Save() error {
 	// Create cookie with the session ID
 	s.setCookie()
 
-	// release session to pool to be re-used on next request
+	// Release session
+	// TODO: It's not safe to use the Session after called Save()
 	releaseSession(s)
 
 	return nil
